+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - The preloader can now be gracefully stopped by sending an interrupt signal to the Sparv process.
 - Added `HeaderAnnotations` and `HeaderAnnotationsAllSourceFiles` classes.
 - Added `korp.keep_undefined_annotations` setting, to include even undefined annotations in the Korp config.
+- Added `dateformat.pre_regex` setting.
 
 ### Changed